Made onedim::first_two_passes work
with vectors of size 2 and implemented SseDouble.

+++ testing.cpp
@@ -27,7 +27,7 @@
 
 static constexpr Int maxdim = 64;
 static constexpr Int chunk_size = 0;
-using ElementType = float;
+using ElementType = double;
 
 using std::chrono::high_resolution_clock;
